requirements.txt
# Các thư viện chính cho RAG Chatbot
flask==3.1.1
python-dotenv==1.1.0
sentence-transformers==4.1.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-openai==0.3.17
faiss-cpu==1.11.0
openai==1.79.0
tiktoken==0.9.0

# Xử lý dữ liệu
numpy==2.2.6
pandas==2.2.3
scikit-learn==1.6.1
tqdm==4.67.1

# Giao diện người dùng (optional)
gradio==5.29.1
pydantic==2.11.4

# Tiện ích và phụ thuộc
requests==2.32.3
python-multipart==0.0.20
aiohttp==3.11.18
transformers==4.51.3
torch==2.7.0